requirements.txt
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
asyncer==0.0.2
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bidict==0.23.1
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
chainlit==1.0.506
charset-normalizer==3.3.2
chevron==0.14.0
click==8.1.7
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.5.14
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.110.3
fastapi-socketio==0.0.10
fastjsonschema==2.19.1
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.4.1
googleapis-common-protos==1.63.0
grpcio==1.63.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib-metadata==7.0.0
ipykernel==6.29.4
ipython==8.24.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.25
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.8
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.1
jupyterlab_widgets==3.0.10
Lazify==0.4.0
literalai==0.0.509
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib-inline==0.1.7
mistune==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
notebook==7.1.3
notebook_shim==0.2.4
openai==1.25.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-exporter-otlp-proto-http==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
overrides==7.7.0
packaging==23.2
pandocfilters==1.5.1
parso==0.8.4
platformdirs==4.2.1
prometheus_client==0.20.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
pure-eval==0.2.2
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-graphql-client==0.4.3
python-json-logger==2.0.7
python-multipart==0.0.9
python-socketio==5.11.2
pywin32==306
pywinpty==2.0.13
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
Send2Trash==1.8.3
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
stack-data==0.6.3
starlette==0.37.2
syncer==2.0.3
terminado==0.18.1
tinycss2==1.3.0
tomli==2.0.1
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
typing_extensions==4.11.0
uptrace==1.24.0
uri-template==1.3.0
urllib3==2.2.1
uvicorn==0.25.0
watchfiles==0.20.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
widgetsnbextension==4.0.10
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.18.1
